show edit and delete for user using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Models\Question;
 use App\Http\Requests\AskQuestionRequest;
 use Illuminate\Console\View\Components\Ask;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -58,7 +59,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied');
+        }
+       return view('questions.edit', compact('question'));
     }
 
     /**
@@ -66,6 +70,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+         if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied');
+        }
         $question->update([
             'title' => $request->title,
             'body' => $request->body
@@ -78,6 +85,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+         if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access Denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
     }
